(original rev 952) 'S' and 'W' now increment and decrement in single slices.
git-svn-id: file:///tmp/current/omni/branches/staging@733 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/view2d/omView2d.h
@@ -90,7 +90,7 @@ protected:
 	void TextureDraw(vector <Drawable*> &textures);
 	void safeTexture(shared_ptr<OmTextureID> gotten_id);
 	void safeDraw(float zoomFactor, int x, int y, int tileLength, shared_ptr<OmTextureID> gotten_id);
-	void PreDraw(Vector2i);
+	void PreDraw(Vector2f);
 
 	void DrawFromFilter (OmFilter2d&);
 	void DrawFromCache();
